Ensure all dependencies are required correctly
The coffee-rails dependency at least was defined in gemspec but never included. Because these aren't specified in a Gemfile anywhere were Rails can use a Bundler.require, we need to ensure all of the dependencies are required manually. All I've done here is copy and paste the dependencies in gemspec and manipulated it into required.

Also, cancan was required, but not present in the dependencies